Updated deleteUser() to make it not possible to delete admin-account

+++ src/Hapi/SQLMethods/Methods.java
@@ -139,6 +139,10 @@ public class Methods {
     public static boolean deleteUser(String username) {
         boolean ok = false;
 
+        if (username.equals("admin")) {
+            return false;
+        }
+
         try {
             String deleteSQL = "DELETE FROM employee WHERE username = ?";
             stm = con.prepareStatement(deleteSQL);
